requirements.txt
torch==1.11.0
torchvision==0.12.0
gradio==2.9.4
transformers==4.20.1
lightning
Pillow==9.2.0
numpy==1.21.6
TTS==0.7.1